[cncwiz] lean on tinyfiledialogs@v3.6.3 for gcode loading
Instead of building a file browser from scratch, let's defer
to the OS to handle this sort of thing. There is one major
caveat though.. while the modal is up, the app doesn't get
normal events and cannot hot reload.

I'm pretty sure this doesn't affect a user's experience, but it
makes the abstraction and way of thinking about hot reload a bit
weird for a developer. Fortunately, there is an escape hatch..
don't use tinyfd in your program if you want to hot reload while
files are being selected. A cimgui based solution would allow for
this.