[gcode/parser] begin handling combinations of commands
This includes things like motion mixed with modal commands. The
main reason for all of this effort is for restarting program
sending. I'd like to be able to rely on my parser instead of
querying grbl for `$G` whenver the user hits abort, which would
look something like:

```
user: abort!
cncwiz: feed hold
cncwiz: $G and parse+store results
cncwiz: soft-reset
```

I've been known to rip the usb cable out of the port in the event of
something really weird happening. Without a parser computing the parser
state on the cncwiz side, it would be impossible to resume a job with
any sort of confidence.