[cncwiz] make program abort work consistently
I was having some difficulty understanding why soft resetting grbl
after a feed hold would alarm (3). Turns out it is due to the machine
coming to rest when it has been in motion.

This has been resolved by updating the status `?` polling rate while
in the feed rate mode and waiting for a few ticks of MPos to be the
same (contiguously).