[cncwiz] add a new homing state to handle non-responsive process
When `$h` is sent to Grbl, it begins homing the cnc machine and stops
responding to status queries (`?`). If you trigger home at the exact
wrong moment however, grbl may respond to a status query with an idle
status. At the end of the homing cycle grbl responds to a status query
with a "home" status.

In order to work around this issue, a "homing" state is manually set
(grbl->state->state) which allows the rx parser handler to ignore
status message that do not change the state to HOME when in a HOMING
state.

The end result is the "grbl is homing..." modal shows every time the
home button is clicked and it dissappears when homing is complete.