[cncwiz] track relative moves between lines
This is one of the final state tracking steps before resuming a program
can be considered functional.

After taking a deeper look into gcode arcs I noticed that the semantics
of X/Y/Z are the same for all of the movement commands -- these words
specify the destination. Given that the end_{x,y,z} can be computed for
arcs as it was for lines!