[pull/stream] all values are created via ps_create_value
This also fixes an inconsistency with value data ownership. The rule
here is that the value->data field is owned by the value and unless
a special destroy_fn is configured `free` will be called.

This semantic change broke counter/multiplier because the requirement
of ownership was not clear. These streams now make a copy of the
incoming data.

This does not detract from the zero-copy goal (which is more applicable
in the IO case) as values can still flow data without copy. However,
this is determined on a stream by stream basis.