[rawkit/gpu] updating an ssbo increments its version number
This removes a very surprising behavior where you'd have to
manually set the ssbo's resource_version before it could actually
be used in a shader.

I originally thought it would be clever to do it this way, but after
having run face first into this wall multiple times it is time to
tear it down.