[rawkit/shader] allow usage of rawkit_gpu_buffer_t
Prior to this you had to use a rawkit_gpu_ssbo_t which effectively
requires double or more likely triple the memory usage.

example:

if we had a 1gb ssbo we'd need the following allocations

- 1gb (gpu) for the actual buffer
- 1gb (gpu) for a staging buffer (host accessible)
- 1gb (cpu) for updating the staging buffer

I think at this point SSBO's are deprecated. I've only kept them
to avoid having to fix all of the examples in this commit.