[project/framegraph] move sum graph building and fix the resulting relative pathing error
To keep a nice api for fg->shader(.., { relative_file, relative_file }),
we need a way to pass in the __FILE__ of the caller. Clang has a function
for this (`__builtin_FILE()`).

The main tradeoff here is that framegraph.h cannot be compiled by other compilers
until this is replaced by a working `std::source_location` implementation.

https://docs.microsoft.com/en-us/cpp/overview/visual-cpp-language-conformance?view=msvc-160