[rawkit/worker,jit] allow guests to spawn workers
This api roughly follows the WebWorker API in two major areas

1. Workers are launched by specifying a separate file. I think you can specify
   a blob in web land, but in rawkit, having a separate files means that we
   can build a new jit program that runs inside of a thread. And by effect,
   the program can hot reload independently of the main program.

2. Communication between worker and host is achieved via messages. Since this
   is a c99 api, postMessage doesn't really work here. So instead we have a pair
   of functions `rawkit_worker_send(<worker>)` and `rawkit_worker_recv(<worker>)`
   and when you want a worker to talk to it's host, you provide the `<worker>`
   argument with the result of `rawkit_worker_host()`

The API will change slightly after this commit as we need to default to a copy
on send by default and provide a separate function for sending by reference.
This is important for stack allocated items.

I'll also be adding a bunch of write through caches that create+store vulkan
queues/command pools/etc.. These will be exposed via the rawkit_worker_* api.