[rawkit/shader] split instance create and begin so workers can provide a queue
There has been a bunch of churn lately trying to get workers launching shaders,
and this is the commit that finally makes it all work.

It is not perfect, but it works. In other words, I'm limping in hopefully
the correct direction.

At some point I'll have all of the vulkan primitives wrapped up w/ their
appropate state, but until then we have this monstrosity.