[rawkit/worker,main] add machinery to allow work synchronization
Workers now house a timeline semaphore and an atomic counter that
can be incremented safely from both the main and worker threads.

This semaphore can be used to make the frame wait for worker work
to complete and the act of registering the semaphore will cause its
counter to increment when submitting a frame. This incremented
counter value can then be used to cause a worker queue submission
to wait on the completion of a frame.

Not all actions in a worker need to synchronize with the frame as
that would mostly defeat the purpose of introducing workers in the
first place.

The default behavior of this system is to do nothing.